+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.15.1
   * Added a utility to convert examples in calls to be const. This is important
     for DDC usage, where they aren't optimized away.
+  * Add a codegen_mode flag, which can be either release or debug. In release
+    mode a missing translation throws an exception, in debug mode it returns the
+    original text, which was the previous behavior.
 
 ## 0.15.0
   * Change non-transformer message rewriting to preserve the original message as